refactor(profile settings): refactor code, split into smaller components
